formed notes into HTML, added annotations and structure
these first two chapters of the Red Book have spoken to me with force and clarity, offering illuminating description of (and offering words for) my own recent inner experiences. this text bursts with Light, rings of Truth, and resonates with me from the depths.

i feel compelled to let this text pass over and through me, to shine upon the words and images that reach out and into me, to annotate with insights of my own.

and as it filters through my lens, the Red Book will be transformed into a new digital medium, where hand-drafted manuscript meets hand-crafted display code.

i do not know, and am guided not to reflect upon, the type of impact this may have -- perhaps it will be none at all. my task is to digest, modernize, and shine upon this beautiful work.