added root index and page image; improved the styling
looking at the text in first two manuscript images, it occurred to me that there was too much translated text. the next manuscript page (according to the PDF) matches the missing page, and it is now included in the prologue.

this commit includes several style improvements, more image references within the chapters, and new website index files.