highlighted "Soul and God" chapter
another beautiful chapter, now that i have finally read it carefully. searching for _something_, first finding an image of in others, until realizing that it is my own soul that is missing. the soul that goes along with me, helping me to see the whole in each part. the recursive is the spring of life, the repeating resonance of well-being.

this is not taught, but discovered: by living and thinking to the fullest, by listening for that internal language you do not understand, by accepting the legitimacy of the knowledge of the heart.

our soul, our God, is our opposite, and believe it this serves our perfection through balance and completion. we must build and create and _live fully_ to develop into something more, to give birth to what is to come.

mockery may come from within and without. you must rise above this, in the example of Christ, to accept that you and he are the same, to become a Christ yourself.

> weight of words and wars we carry
> i'm like you
> just like you
> eyes of secret storm and story
> show and tell: we'll make it through
> by the _telling_
> will they _become_?
> will they all be
> feathers?