highlighted "Service of the Soul" chapter
this chapter teaches how to handle the pressures of fear, doubt, and defiance that are felt against the discovery of soul. or rather, it teache how to _resist_ those as temptations while at the _same time_ acknowledging that they are real and justified concerns.

if you can find a way to marry your soul's flood of chaos into the order of the world and your self, you shall produce the divine child -- the supreme meaning.