formatted "Experiences in the Desert"
this chapter begins with dialog with the soul, and ends in a pillar. it spans two manuscript pages.

i have modified my formatting process to include a more careful read-through of the text. there are multiple goals for this: view the bigger-picture meaning of the chapter before highlighting, achieve better initial structuring, catch more OCR errors, and (hopefully) to learn and understand the text more completely.