highlighted "Experiences in the Desert"
this chapter teaches us that the soul is not here for us to take from, to complain at, to give us pleasure. it demands that we _wait_, not expecting anything to fall to us easily, not _measuring_ our soul to death, and certainly not making plans and intentions for any Light we may receive.

your self-scorn, the criticism and disrespect and unworthiness you lash upon, is born of vanity. be a vessel of the Light, burst forth with it. leave your self-scorn behind; these words speak from beyond _you_.

and again, we see the supreme meaning arising from the merging of opposites. find wisdom that is simple, without accepting artificial foolishness. find cleverness that accepts the simple, without mockery. weigh your cleverness against your wisdom. find, _be_, the balance and elegance in the simplicity. be _clever fools._