highlighted the first half of "Descent into Hell"
i saw an idea-image while highlighting (and listening to the _Laturalus_ album, especially _Schism_; also currently reading _Neverwhere_, which keeps hinting at _Ship of Theseus_). this was its path:

it collapses possibilities when i churn on ideas, trying to make sense of every single thing. crystallizing into meaning is murder to the True understanding. understanding is a _bridge_, and the measuring of meaning _burns_ that bridge. i know the pieces fit.

this True understanding is in the connections between the very many things, the similarities, the differences, the relations, the symbols, the images, their contexts and juxtapositions. it is a training of your mind to find the seams and correlations between the real and the nonsense. i know the pieces fit.

between supposed lovers. marry the real and the nonsense. heal them. make them one. learn the non-sense: freedom from meaning in the abstract, full of potential and possibility and _life_. learn to accept a state of non-meaning in your mind. let your pieces drift and join and split. allow yourself to change, to transmutate. to become perceptive in new directions as your pieces enter new configurations. find the ones you like best. grow into them. we know the pieces fit.

---

so, if there cannot be _meaning_ in this understanding-sense, what _is_ there? does it help you find your best _you_? the red book tells me this mixing, this **balance** between our halves, is the **_supreme_ meaning,** or perhaps the _image_ of it.

i do not dare to push further with this measurement. it would seem: the more explanation, the more collapse, the more burning. i do not wish to cause undue damage.

still: i feel a terrible urgency. i feel more bridges will need to burn, and if that is the case, we must strive to collapse wisely and efficiently. traffic safely in art.

> the poetry
> that comes
> from the squaring-off between
> and the circling
> is worth it
> finding beauty in
> the dissonance

squaring-off between supposed lovers
the real and the imaginary
the meaning and the nonsense
the knowing and the `NaN`sense
and circling forever, Theseus
swirls of blackest destruction and creation
such poetry! their squaring-off between
the weight of words and wars
my bridge of understanding
crystallizes without collapse
encoded within expression
delivered to your mind
forms in you ideas
all meaning implied
simply stepped aside
stop burning bridges
balance for the Light