highlighted the second half of "Descent into Hell"
i brought forth a strange plant today: the _tenput_ seed. an image of software rooted in art and rage and faith. WE HAVE SEEN software without soul.

the descent grows darker in this half. take counsel in your fears. be prepared: you must both sacrifice and be sacrificed, inflicting the suffering of the world unto your own heart, knowing it was done by your own heinous hand. new life can only be born of such torment.

events are meaningless, yet we create meaning for them anyway. we do this in part to bequeath those images unto our future. but these events, these things outside of us, this is dead, the already-has-been. but life is _inside_ of us, where the meaning is created.