highlighted the first half of "Splitting the Soul"
indigence and defiance. war. here we rage against soul, and perhaps miss something of what soul is telling us. our lust for battle, our internal scream _to be free_, to defy obstacles in our path; all this blinds us from Truth and Light.

the soul is of a place of infinities. its nature is the being-of-all-things. our worldly minds cannot perceive all things simultaneously; many of us struggle with the seeing of even _one thing_ with clarity. if the soul presents some face, some personality, some archetype to us, this does not equate to the nature of the soul. it is a mask.

in this chapter we believe we rage against _soul,_ yet we merely rage against _mask._

is war against _mask_ worthwhile? perhaps. could it be _civil_ war? could we rage against the _mask_ that controls and subordinates us? could we become _peer_ instead of _serf_? could we stand *boldly,* _unyielding_ against such oppression? could we fight while hating the fight? _wrestle_ and not _kill_? minimize suffering, resolve the dispute, heal, _thrive._

Ender knows of this (and of trickery):

> In the moment when I truly _understand_ my enemy, understand him well enough to defeat him, then in that _very moment_ I also _love_ him. I think it’s impossible to really _understand_ somebody, what they want, what they believe, and not love them the way they love themselves. And then, in that very moment when I love them... _I destroy them._

> I need you to be clever, Bean. I need you to think of solutions to problems we haven't seen yet. I want you to try things that no one has ever tried because they're absolutely stupid.

> The enemy's gate is _down._

> And always Ender carried with him a dry white cocoon, looking for a place where the hive-queen could awaken and thrive in peace. He looked a long time.